waggle-dance-core/src/main/java/com/hotels/bdp/waggledance/client/MetaStoreUriNormaliser.java
package com.hotels.bdp.waggledance.client;

import java.net.URI;
import java.net.URISyntaxException;
import java.util.Set;
import java.util.TreeSet;

import com.google.common.base.Joiner;

final class MetaStoreUriNormaliser {

  private MetaStoreUriNormaliser() {};

  static String normaliseMetaStoreUris(String metaStoreUris) {
    try {
      String[] rawUris = metaStoreUris.split(",");
      Set<String> uris = new TreeSet<>();
      for (String rawUri : rawUris) {
        URI uri = new URI(rawUri);
        uris.add(uri.toString());
      }
      return Joiner.on(",").join(uris);
    } catch (URISyntaxException e) {
      throw new RuntimeException(e);
    }
  }

}
